requirements.txt
# Core FastAPI dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP client libraries
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# Web scraping
beautifulsoup4>=4.12.0

# Aviation weather data
avwx-engine>=1.8.0

# Environment management
python-dotenv>=1.0.0

# Network utilities
urllib3>=2.0.0

# ── v3 additions ─────────────────────────────────────────

# ML / Data Science
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Caching
redis>=5.0.0

# FAA SWIM (Solace messaging for NOTAMs)
solace-pubsubplus>=1.8.0

# Date/Time utilities
python-dateutil>=2.8.0
pytz>=2023.3
